changes to accept any file type

+++ index.js
@@ -1,14 +1,15 @@
 const fs = require('fs');
-const glob = require('glob');
 const path = require('path');
 
 module.exports = {
-  onPostBuild: ({ constants: { PUBLISH_DIR }, utils }) => {
-    console.log('🚮 Removing all the JS files.');
+  onPostBuild: ({ constants: { PUBLISH_DIR }, inputs, utils }) => {
+    if (inputs.fileType == '') return console.log('🤦🏻‍♀️ Oops, need a file type.');
+    console.log(`🚮 Removing all the ${inputs.fileType} files.`);
 
+    const fileTypeLength = inputs.fileType.length;
     const files = fs
       .readdirSync(PUBLISH_DIR)
-      .filter((file) => file.slice(-2) == 'js');
+      .filter((file) => file.slice(-fileTypeLength) == inputs.fileType);
 
     files.forEach((file) => {
       fs.unlinkSync(`${PUBLISH_DIR}/${file}`, (error) => {
@@ -16,8 +17,5 @@ module.exports = {
       });
       console.log(`${file} has been deleted 💣`);
     });
-
-    const builtPages = glob.sync(`${PUBLISH_DIR}/*`);
-    console.log(builtPages);
   },
 };
